+++ spec/rubocop/cop/rspec/nested_groups_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 RSpec.describe RuboCop::Cop::RSpec::NestedGroups, :config do
-  it 'flags nested contexts' do
+  it 'flags nested example groups defined inside `describe`' do
     expect_offense(<<-RUBY)
       describe MyClass do
         context 'when foo' do
@@ -14,6 +14,36 @@ RSpec.describe RuboCop::Cop::RSpec::NestedGroups, :config do
 
         context 'when qux' do
           context 'when norf' do
+          end
+        end
+      end
+    RUBY
+  end
+
+  it 'flags nested example groups' do
+    expect_offense(<<-RUBY)
+      example_group MyClass do
+        context 'when foo' do
+          context 'when bar' do
+            context 'when baz' do
+            ^^^^^^^^^^^^^^^^^^ Maximum example group nesting exceeded [4/3].
+            end
+          end
+        end
+      end
+    RUBY
+  end
+
+  it 'flags nested example groups inside shared examples' do
+    expect_offense(<<-RUBY)
+      shared_examples_for 'nested like express' do
+        context 'when foo' do
+          context 'when bar' do
+            context 'when baz' do
+              context 'when qux' do
+              ^^^^^^^^^^^^^^^^^^ Maximum example group nesting exceeded [4/3].
+              end
+            end
           end
         end
       end
@@ -35,11 +65,32 @@ RSpec.describe RuboCop::Cop::RSpec::NestedGroups, :config do
     expect(cop.config_to_allow_offenses[:exclude_limit]).to eq('Max' => 4)
   end
 
-  it 'ignores non-spec context methods' do
-    expect_no_offenses(<<-RUBY)
+  it 'flags example groups wrapped in classes' do
+    expect_offense(<<-RUBY)
       class MyThingy
-        context 'this is not rspec' do
-          context 'but it uses contexts' do
+        describe MyClass do
+          context 'when foo' do
+            context 'when bar' do
+              context 'when baz' do
+              ^^^^^^^^^^^^^^^^^^ Maximum example group nesting exceeded [4/3].
+              end
+            end
+          end
+        end
+      end
+    RUBY
+  end
+
+  it 'flags example groups wrapped in modules' do
+    expect_offense(<<-RUBY)
+      module MyNamespace
+        describe MyClass do
+          context 'when foo' do
+            context 'when bar' do
+              context 'when baz' do
+              ^^^^^^^^^^^^^^^^^^ Maximum example group nesting exceeded [4/3].
+              end
+            end
           end
         end
       end